pyproject.toml
[tool.poetry]
name = "stream-tool"
version = "0.1.0"
description = ""
authors = ["Skyler Ficklin <skylerficklin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
obs-websocket-py = "^1.0"
flask = "^3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
